feat(commit): Add `message()` method and `MessageRef` type…(#198)
…allowing a better understanding of typical commit messages.
